6/11
Today we wanted to create a menu as to make it possible to create objects on which we could place building blocks. We were stopped in our tracks by the fact that we could not get the script to recognize our SteamVR button input. As such we have just created several of the TowerFoundation objects and placed them outside the castle to demonstrate how the building works so far.

This concludes this weeks goal of creating a crafting system, but does not fullfill the entire vision. As such what is to be done this week is as follows:

Questions for Ebbe:
How can we create a menu?
(floating Pane, scrolling through options)

How can we obtain input from the controllers?

How can we make foundations that can be placed so the towers wont just float mid air after creation

To be done:

Create menu after talking with Ebbe.

Create towerFoundations that can be moved by the player.