14 / 11 / 18
Today we continued on the building menu. Since we dont have a pointer that can generate an actual collision point with the ground, we tried to create a script to fire a gameObject outwards, make it collide with the board and then take that Objects location and transform that into where we would place a foundation / ball.  We also tried to create a laser with the RayCast object, but the raycast doesnt seem to work in the slightest.

The solution seems to be in the telepot script, we will scour that and the internet for a solution til monday.

We also started on a pathing system, which is in the early stages.